process-images/requirements.txt
google.cloud
google.cloud.storage
google.cloud.documentai
google.cloud.pubsub
google.api_core
functions-framework==3.2.0